+++ goleft.go
@@ -1,3 +1,3 @@
 package goleft
 
-const Version = "0.1.13"
+const Version = "0.1.14-dev"